Fixed a bug in DeepLesion preprocessing

+++ data/deep_lesion/download_deep_lesion.sh
@@ -15,7 +15,11 @@ num_files=${#files[@]}
 for (( i=0; i<${num_files}; i++ ));
 do
     idx=$(printf "%02d" "$(expr $i + 1)")
-    wget ${files[$i]} -O Images_png_$idx.zip
-    unzip Images_png_$idx.zip
+    if [ -f Images_png_$idx.zip ]; then
+        echo "Images_png_$idx.zip exist"
+    else
+        wget ${files[$i]} -O Images_png_$idx.zip
+        unzip Images_png_$idx.zip
+    fi
 done
 echo "DeepLesion dataset downloaded!"
